Aeval: CLI code refactor
Move code for parsing command line arguments (getIntValue, etc.) into a
dedicated header file for easier reuse.
