+++ tools/aeval/Ae.cpp
@@ -1,5 +1,6 @@
 #include "ae/AeValSolver.hpp"
 #include "ufo/Smt/EZ3.hh"
+#include "ae/CLIParsing.hpp"
 
 using namespace ufo;
 
@@ -20,45 +21,6 @@ using namespace ufo;
  *   ../test/ae/example1_t_part.smt2
  *
  */
-
-bool getBoolValue(const char * opt, bool defValue, int argc, char ** argv)
-{
-  for (int i = 1; i < argc; i++)
-  {
-    if (strcmp(argv[i], opt) == 0) return true;
-  }
-  return defValue;
-}
-
-int getIntValue(const char * opt, int defValue, int argc, char ** argv)
-{
-  for (int i = 1; i < argc-1; i++)
-  {
-    if (strcmp(argv[i], opt) == 0)
-    {
-      char* p;
-      int num = strtol(argv[i+1], &p, 10);
-      if (*p) return 1;      // if used w/o arg, return boolean
-      else return num;
-    }
-  }
-  return defValue;
-}
-
-char * getSmtFileName(int num, int argc, char ** argv)
-{
-  int num1 = 1;
-  for (int i = 1; i < argc; i++)
-  {
-    int len = strlen(argv[i]);
-    if (len >= 5 && strcmp(argv[i] + len - 5, ".smt2") == 0)
-    {
-      if (num1 == num) return argv[i];
-      else num1++;
-    }
-  }
-  return NULL;
-}
 
 void printUsage()
 {
